requirements.txt
Django
coverage
djangorestframework<4
py
pytest
